added tests and fixed serverless and travis
