+++ .travis.yml
@@ -5,8 +5,8 @@ rust:
 - beta
 - nightly
 
-#services:
-  #- docker
+services:
+  - docker
   
 before_install:
 - wget -qO- https://raw.githubusercontent.com/creationix/nvm/v0.33.11/install.sh | bash
@@ -23,8 +23,9 @@ script:
   - cargo clean
   - cargo build --release --target=x86_64-unknown-linux-musl
   - cargo test
-  #- sudo docker pull lambci/lambda:go1.x  #this is to download the docker image if it doesnt exist yet
+  - sudo docker pull lambci/lambda:provided  #this is to download the docker image if it doesnt exist yet
   - npm install
+  - node index
   - npm test
 
 after_success: |
