+++ tests/parameter3.json
@@ -6,5 +6,48 @@
         "sensitivity":"price",
         "optionType":"put"
     },
-    "queryParameters":{}
+    "queryStringParameters":{},
+    "httpMethod":"POST",
+    "headers": {
+        "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
+        "Accept-Encoding": "gzip, deflate, lzma, sdch, br",
+        "Accept-Language": "en-US,en;q=0.8",
+        "CloudFront-Forwarded-Proto": "https",
+        "CloudFront-Is-Desktop-Viewer": "true",
+        "CloudFront-Is-Mobile-Viewer": "false",
+        "CloudFront-Is-SmartTV-Viewer": "false",
+        "CloudFront-Is-Tablet-Viewer": "false",
+        "CloudFront-Viewer-Country": "US",
+        "Host": "wt6mne2s9k.execute-api.us-west-2.amazonaws.com",
+        "Upgrade-Insecure-Requests": "1",
+        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/52.0.2743.82 Safari/537.36 OPR/39.0.2256.48",
+        "Via": "1.1 fb7cca60f0ecd82ce07790c9c5eef16c.cloudfront.net (CloudFront)",
+        "X-Amz-Cf-Id": "nBsWBOrSHMgnaROZJK1wGCZ9PcRcSpq_oSXZNQwQ10OTZL4cimZo3g==",
+        "X-Forwarded-For": "192.168.100.1, 192.168.1.1",
+        "X-Forwarded-Port": "443",
+        "X-Forwarded-Proto": "https"
+    },
+    "stageVariables": {"stageVarName": "stageVarValue"},
+    "requestContext": {
+        "resourceId": "123456",
+        "apiId": "1234567890",
+        "resourcePath": "/",
+        "httpMethod": "GET",
+        "requestId": "c6af9ac6-7b61-11e6-9a41-93e8deadbeef",
+        "accountId": "123456789012",
+        "identity": {
+          "apiKey": null,
+          "userArn": null,
+          "cognitoAuthenticationType": null,
+          "caller": null,
+          "userAgent": "Custom User Agent String",
+          "user": null,
+          "cognitoIdentityPoolId": null,
+          "cognitoIdentityId": null,
+          "cognitoAuthenticationProvider": null,
+          "sourceIp": "127.0.0.1",
+          "accountId": null
+        },
+        "stage": "prod"
+    }
 }
